+++ src/main.rs
@@ -35,7 +35,7 @@ fn benchmark_encode(iterations    : usize,
                     pparam        : ParallelParam) {
     let mut shards = make_random_shards!(per_shard, data_shards + parity_shards);
     //let mut shards = make_blank_shards(per_shard, data_shards + parity_shards);
-    let r = ReedSolomon::with_pparam(data_shards, parity_shards, pparam);
+    let r = ReedSolomon::with_pparam(data_shards, parity_shards, pparam).unwrap();
 
     let start = time::precise_time_ns();
     for _ in 0..iterations {
@@ -66,7 +66,7 @@ fn benchmark_encode_inplace(iterations    : usize,
     //let mut shards = make_blank_shards(per_shard, data_shards + parity_shards);
     let mut slices : [[u8; PER_SHARD]; DATA_SHARDS + PARITY_SHARDS] =
         [[0; PER_SHARD]; DATA_SHARDS + PARITY_SHARDS];
-    let r = ReedSolomon::with_pparam(DATA_SHARDS, PARITY_SHARDS, pparam);
+    let r = ReedSolomon::with_pparam(DATA_SHARDS, PARITY_SHARDS, pparam).unwrap();
 
     let mut slices_ref : Vec<&mut [u8]> =
         Vec::with_capacity(DATA_SHARDS + PARITY_SHARDS);
@@ -98,7 +98,7 @@ fn benchmark_verify(iterations    : usize,
                     pparam        : ParallelParam) {
     let mut shards = make_random_shards!(per_shard, data_shards + parity_shards);
     //let mut shards = make_blank_shards(per_shard, data_shards + parity_shards);
-    let r = ReedSolomon::with_pparam(data_shards, parity_shards, pparam);
+    let r = ReedSolomon::with_pparam(data_shards, parity_shards, pparam).unwrap();
 
     r.encode_shards(&mut shards).unwrap();
 
@@ -125,7 +125,7 @@ fn benchmark_reconstruct(iterations    : usize,
                          pparam        : ParallelParam) {
     let mut shards = make_random_shards!(per_shard, data_shards + parity_shards);
     //let mut shards = make_blank_shards(per_shard, data_shards + parity_shards);
-    let r = ReedSolomon::with_pparam(data_shards, parity_shards, pparam);
+    let r = ReedSolomon::with_pparam(data_shards, parity_shards, pparam).unwrap();
 
     r.encode_shards(&mut shards).unwrap();