+++ reed-solomon-erasure/CHANGELOG.md
@@ -1,3 +1,29 @@
+## 2.3.2
+  - Made `==`(`eq`) for `ReedSolomon` more reasonable
+    - Previously `==` would compare
+      - data shard count
+      - parity shard count
+      - total shard count
+      - internal encoding matrix
+      - internal `ParallelParam`
+    - Now it only compares
+      - data shard count
+      - parity shard count
+
+## 2.3.1
+  - Added info on encoding behaviour to doc
+
+## 2.3.0
+  - Made Reed-Solomon codec creation methods return error instead of panic when shard numbers are not correct
+
+## 2.2.0
+  - Fixed SBS error checking code
+  - Documentation fixes and polishing
+  - Renamed `Error::InvalidShardsIndicator` to `Error::InvalidShardFlags`
+  - Added more details to documentation on error handling
+  - Error handling code overhaul and checks for all method variants
+  - Dead commented out code cleanup and indent fix
+
 ## 2.1.0
   - Added Nicolas's SIMD C code files, gaining major speedup on supported CPUs
   - Added support for "shard by shard" encoding, allowing easier streamed encoding